Switch from using g_utf8_strlen to strlen
This fixes an issue where messages with non-ASCII characters get cut off partway through, either invalidating the entire message, or only sending part of the characters in the message.

This is an issue inherited from the original project skeleton --- g_utf8_strlen counts the number of *characrters*, but we only care about *bytes*. This matters when characters can take up more than 1 byte.